+++ src/main/java/raisetech/StudentManagement/StudentManagementApplication.java
@@ -1,32 +1,14 @@
 package raisetech.StudentManagement;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 
 @SpringBootApplication
-@RestController
 public class StudentManagementApplication {
-
-    @Autowired
-    private StudentRepository repository;
 
 	public static void main(String[] args) {
 		SpringApplication.run(StudentManagementApplication.class, args);
 	}
-
-	@GetMapping("/student")
-	public List<Student> getStudentList() {
-		return repository.search();
-	}
-
-	@GetMapping("/studentsCourseList")
-	public List<StudentsCourses> getStudentsCourseList() {
-		return repository.sarchStudentCourses();
-	}
-
 
 }
