+++ src/main/java/raisetech/StudentManagement/controller/StudentController.java
@@ -1,0 +1,31 @@
+package raisetech.StudentManagement.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentsCourses;
+import raisetech.StudentManagement.service.StudentService;
+
+import java.util.List;
+
+@RestController
+public class StudentController {
+
+    private StudentService service;
+
+    @Autowired
+    public StudentController(StudentService service) {
+        this.service = service;
+    }
+
+    @GetMapping("/student")
+    public List<Student> getStudentList() {
+        return service.searchStudentList();
+    }
+
+    @GetMapping("/studentsCourseList")
+    public List<StudentsCourses> getStudentsCourseList() {
+        return service.searchStudentsCourseList();
+    }
+}
