+++ src/main/java/raisetech/StudentManagement/data/StudentsCourses.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement;
+package raisetech.StudentManagement.data;
 
 
 import lombok.Getter;
